+++ .gitignore
@@ -367,3 +367,6 @@ obj/
 /packages/
 riderModule.iml
 /_ReSharper.Caches/
+
+appsettings.json
+appsettings.Development.json
