+++ UserModule.Contracts/Repositories/IRoleRepository.cs
@@ -11,8 +11,8 @@ namespace UserModule.Contracts.Repositories
 {
     public interface IRoleRepository : IBaseEntityRepository<Role>
     {
-        public Task<List<Role>> GetRoles(List<RoleName> roleNames);
+        public Task<List<Role>> GetRolesAsync(List<RoleName> roleNames);
 
-        public Task<List<Role>> GetRolesByUserId(Guid userId);
+        public Task<List<Role>> GetRolesByUserIdAsync(Guid userId);
     }
 }
